refactor : modify output statement
